Removed doctypemigrations django app
